test commit

+++ test/test_commit.py
@@ -1,11 +1,11 @@
 import sys
 
 
-def addition(a: float, b: float):
+def addition(a: int, b: int):
     return a + b
 
 
 if __name__ == "__main__":
     num1 = sys.argv[0]
     num2 = sys.argv[1]
-    print(addition(float(num1), float(num2)))
+    print(addition(int(num1), int(num2)))
